Move RouteContract into Contracts namespace
