+++ packages/framework/src/Models/NavigationMenu.php
@@ -2,8 +2,8 @@
 
 namespace Hyde\Framework\Models;
 
+use Hyde\Framework\Contracts\RouteContract;
 use Hyde\Framework\Modules\Routing\Route;
-use Hyde\Framework\Modules\Routing\RouteContract;
 use Hyde\Framework\Modules\Routing\Router;
 use Illuminate\Support\Collection;
