+++ packages/framework/src/Modules/Routing/Route.php
@@ -3,6 +3,7 @@
 namespace Hyde\Framework\Modules\Routing;
 
 use Hyde\Framework\Contracts\PageContract;
+use Hyde\Framework\Contracts\RouteContract;
 use Hyde\Framework\Hyde;
 use Illuminate\Support\Collection;
