+++ packages/framework/src/Modules/Routing/RouteFacadeContract.php
@@ -3,12 +3,13 @@
 namespace Hyde\Framework\Modules\Routing;
 
 use Hyde\Framework\Contracts\PageContract;
+use Hyde\Framework\Contracts\RouteContract;
 use Illuminate\Support\Collection;
 
 /**
  * This contract defines the static facade methods for the Route class.
  *
- * @see \Hyde\Framework\Modules\Routing\RouteContract for the interface that each route model must implement.
+ * @see \Hyde\Framework\Contracts\RouteContract for the interface that each route model must implement.
  */
 interface RouteFacadeContract
 {
@@ -18,7 +19,7 @@ interface RouteFacadeContract
      * Alias for static::getFromKey().
      *
      * @param  string  $routeKey  Example: posts/foo.md
-     * @return \Hyde\Framework\Modules\Routing\RouteContract
+     * @return \Hyde\Framework\Contracts\RouteContract
      *
      * @throws \Hyde\Framework\Modules\Routing\RouteNotFoundException
      */
@@ -28,7 +29,7 @@ interface RouteFacadeContract
      * Get a route from the Router index for the specified route key.
      *
      * @param  string  $routeKey  Example: posts/foo.md
-     * @return \Hyde\Framework\Modules\Routing\RouteContract
+     * @return \Hyde\Framework\Contracts\RouteContract
      *
      * @throws \Hyde\Framework\Modules\Routing\RouteNotFoundException
      */
@@ -38,7 +39,7 @@ interface RouteFacadeContract
      * Get a route from the Router index for the specified source file path.
      *
      * @param  string  $sourceFilePath  Example: _posts/foo.md
-     * @return \Hyde\Framework\Modules\Routing\RouteContract
+     * @return \Hyde\Framework\Contracts\RouteContract
      *
      * @throws \Hyde\Framework\Modules\Routing\RouteNotFoundException
      */
@@ -48,7 +49,7 @@ interface RouteFacadeContract
      * Get a route from the Router index for the supplied page model.
      *
      * @param  \Hyde\Framework\Contracts\PageContract  $page
-     * @return \Hyde\Framework\Modules\Routing\RouteContract
+     * @return \Hyde\Framework\Contracts\RouteContract
      *
      * @throws \Hyde\Framework\Modules\Routing\RouteNotFoundException
      */
@@ -57,7 +58,7 @@ interface RouteFacadeContract
     /**
      * Get all routes from the Router index.
      *
-     * @return \Illuminate\Support\Collection<\Hyde\Framework\Modules\Routing\RouteContract>
+     * @return \Illuminate\Support\Collection<\Hyde\Framework\Contracts\RouteContract>
      */
     public static function all(): Collection;
 }
