+++ packages/framework/src/Modules/Routing/Router.php
@@ -3,6 +3,7 @@
 namespace Hyde\Framework\Modules\Routing;
 
 use Hyde\Framework\Contracts\PageContract;
+use Hyde\Framework\Contracts\RouteContract;
 use Hyde\Framework\Contracts\RouterContract;
 use Hyde\Framework\Helpers\Features;
 use Hyde\Framework\Models\Pages\BladePage;
@@ -43,7 +44,7 @@ class Router implements RouterContract
     /**
      * The routes discovered by the router.
      *
-     * @var \Illuminate\Support\Collection<\Hyde\Framework\Modules\Routing\RouteContract>
+     * @var \Illuminate\Support\Collection<\Hyde\Framework\Contracts\RouteContract>
      */
     protected Collection $routes;
