+++ packages/framework/src/Services/BuildService.php
@@ -3,8 +3,8 @@
 namespace Hyde\Framework\Services;
 
 use Hyde\Framework\Concerns\InteractsWithDirectories;
+use Hyde\Framework\Contracts\RouteContract as Route;
 use Hyde\Framework\Hyde;
-use Hyde\Framework\Modules\Routing\RouteContract as Route;
 use Hyde\Framework\Modules\Routing\Router;
 use Hyde\Framework\StaticPageBuilder;
 use Illuminate\Console\Concerns\InteractsWithIO;
