+++ packages/framework/tests/Feature/RouteTest.php
@@ -2,13 +2,13 @@
 
 namespace Hyde\Framework\Testing\Feature;
 
+use Hyde\Framework\Contracts\RouteContract;
 use Hyde\Framework\Hyde;
 use Hyde\Framework\Models\Pages\BladePage;
 use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\Pages\MarkdownPost;
 use Hyde\Framework\Modules\Routing\Route;
-use Hyde\Framework\Modules\Routing\RouteContract;
 use Hyde\Framework\Modules\Routing\RouteNotFoundException;
 use Hyde\Framework\Modules\Routing\Router;
 use Hyde\Testing\TestCase;
